Group digital take-up tests thematically into one file
